+++ src/component/toolbox_position.rs
@@ -4,7 +4,7 @@ use amethyst::ecs::{Component, DenseVecStorage};
 #[derive(Debug)]
 pub struct ToolboxPosition {
     pub toolbox: usize,
-    pub index: usize
+    pub index: i32
 }
 
 impl Component for ToolboxPosition {
